Standardise online client disconnect handling
