+++ osu.Game/Online/Spectator/SpectatorClient.cs
@@ -76,11 +76,6 @@ namespace osu.Game.Online.Spectator
         public event Action<int, long>? OnUserScoreProcessed;
 
         /// <summary>
-        /// Invoked just prior to disconnection requested by the server via <see cref="IStatefulUserHubClient.DisconnectRequested"/>.
-        /// </summary>
-        public event Action? Disconnecting;
-
-        /// <summary>
         /// A dictionary containing all users currently being watched, with the number of watching components for each user.
         /// </summary>
         private readonly Dictionary<int, int> watchedUsersRefCounts = new Dictionary<int, int>();
@@ -200,18 +195,6 @@ namespace osu.Game.Online.Spectator
                 watchingUsers.RemoveAll(u => u.OnlineID == userId);
             });
 
-            return Task.CompletedTask;
-        }
-
-        Task IStatefulUserHubClient.DisconnectRequested()
-        {
-            Schedule(() => DisconnectInternal().FireAndForget());
-            return Task.CompletedTask;
-        }
-
-        Task IStatefulUserHubClient.ServerShuttingDown()
-        {
-            // TODO: check when we can disconnect and do so.
             return Task.CompletedTask;
         }
 
@@ -379,12 +362,6 @@ namespace osu.Game.Online.Spectator
 
         protected abstract Task StopWatchingUserInternal(int userId);
 
-        protected virtual Task DisconnectInternal()
-        {
-            Disconnecting?.Invoke();
-            return Task.CompletedTask;
-        }
-
         protected override void Update()
         {
             base.Update();
@@ -452,5 +429,32 @@ namespace osu.Game.Online.Spectator
                 });
             });
         }
+
+        #region Disconnection handling
+
+        /// <summary>
+        /// Invoked just prior to disconnection.
+        /// </summary>
+        public event Action? Disconnecting;
+
+        protected abstract Task DisconnectInternal();
+
+        Task IStatefulUserHubClient.DisconnectRequested()
+        {
+            Schedule(() =>
+            {
+                Disconnecting?.Invoke();
+                DisconnectInternal().FireAndForget();
+            });
+            return Task.CompletedTask;
+        }
+
+        Task IStatefulUserHubClient.ServerShuttingDown()
+        {
+            // TODO: check when we can disconnect and do so.
+            return Task.CompletedTask;
+        }
+
+        #endregion
     }
 }
