+++ osu.Game/Tests/Visual/Multiplayer/TestMultiplayerClient.cs
@@ -824,9 +824,9 @@ namespace osu.Game.Tests.Visual.Multiplayer
             return MessagePackSerializer.Deserialize<T>(serialized, SignalRUnionWorkaroundResolver.OPTIONS);
         }
 
-        public override Task DisconnectInternal()
+        protected override Task DisconnectInternal()
         {
-            isConnected.Value = false;
+            Disconnect();
             return Task.CompletedTask;
         }
