+++ osu.Game/Tests/Visual/Spectator/TestSpectatorClient.cs
@@ -204,10 +204,10 @@ namespace osu.Game.Tests.Visual.Spectator
             });
         }
 
-        protected override async Task DisconnectInternal()
+        protected override Task DisconnectInternal()
         {
-            await base.DisconnectInternal().ConfigureAwait(false);
             isConnected.Value = false;
+            return Task.CompletedTask;
         }
     }
 }
